update product editing

+++ lib/ui/screens/admin/list.dart
@@ -9,38 +9,42 @@ class ProductListPage extends StatelessWidget {
     return BlocBuilder(
       bloc: bloc,
       builder: (context, state) {
-        return ListView.builder(
-          itemBuilder: (context, index) {
-            var product = state.products[index];
-            handleOpenEdit() => bloc.dispatch(SelectProduct(product.id));
+        if (state is ProductReady) {
+          return ListView.builder(
+            itemBuilder: (context, index) {
+              var product = state.products[index];
+              handleOpenEdit() => bloc.dispatch(SelectProduct(product.id));
 
-            return Dismissible(
-              key: Key(product.id),
-              onDismissed: (direction) {
-                bloc.dispatch(DeleteProduct(product.id));
-              },
-              background: Container(color: Colors.red),
-              child: Column(
-                children: <Widget>[
-                  ListTile(
-                    leading: CircleAvatar(
-                      backgroundImage: NetworkImage(product.image),
+              return Dismissible(
+                key: Key(product.id),
+                onDismissed: (direction) {
+                  bloc.dispatch(DeleteProduct(product.id));
+                },
+                background: Container(color: Colors.red),
+                child: Column(
+                  children: <Widget>[
+                    ListTile(
+                      leading: CircleAvatar(
+                        backgroundImage: NetworkImage(product.image),
+                      ),
+                      title: Text(product.title),
+                      subtitle: Text(product.price.toString()),
+                      trailing: _buildEditButton(
+                        context,
+                        product,
+                        handleOpenEdit,
+                      ),
                     ),
-                    title: Text(product.title),
-                    subtitle: Text('\$${product.price.toString()}'),
-                    trailing: _buildEditButton(
-                      context,
-                      product,
-                      handleOpenEdit,
-                    ),
-                  ),
-                  Divider()
-                ],
-              ),
-            );
-          },
-          itemCount: state.products.length,
-        );
+                    Divider()
+                  ],
+                ),
+              );
+            },
+            itemCount: state.products.length,
+          );
+        } else {
+          return Container();
+        }
       },
     );
   }
